services/processor/pyproject.toml
[tool.poetry]
name = "processor"
version = "0.1.0"
description = "Document processing service for Personal RAG-LLM"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
pdfplumber = "^0.10.1"
pymupdf = "^1.23.8"
pytesseract = "^0.3.10"
extract-msg = "^0.45.0"
python-docx = "^1.1.0"
sentence-transformers = "^2.5.0"
neo4j = "^5.13.0"
chromadb = "^0.4.22"
duckdb = "^0.9.2"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
requests = "^2.31.0"
pillow = "^10.0.0"
aiohttp = "^3.8.0"
ollama = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pylint = "^3.0.2"
black = "^23.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
